Added the media for screen size when it's smaller than 600px
Added a bunch of fixes to the styling of the page when the size of the screen is smaller than 600px. This also affects the sidebar width.

Made the grid display have one column instead of the two so that content fits, and also aligns all the text to the center.

Changed the font size and padding to all of the input and labels in the "FormPara" field, aka the form.
Also gave them all new widths and Max-Widths as well.

Added a "maxlength" to the textarea, to prevent too long of an explanation.